+++ lib/bolt/outputter/json.rb
@@ -36,7 +36,7 @@ module Bolt
         @stream.puts "],\n"
         @preceding_item = false
         @items_open = false
-        @stream.puts format('"node_count": %<size>d, "elapsed_time": %<elapsed>d }',
+        @stream.puts format('"target_count": %<size>d, "elapsed_time": %<elapsed>d }',
                             size: results.size,
                             elapsed: elapsed_time)
       end