+++ modules/aggregate/lib/puppet/functions/aggregate/count.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 # Aggregates the key/value pairs in the results of a ResultSet into a hash
-# mapping the keys to a hash of each distinct value and how many nodes returned
+# mapping the keys to a hash of each distinct value and how many targets returned
 # that value for the key.
 Puppet::Functions.create_function(:'aggregate::count') do
   dispatch :aggregate_count do