+++ modules/puppetdb_fact/spec/plans/init_spec.rb
@@ -17,13 +17,13 @@ describe 'puppetdb_fact' do
 
   it 'returns facts from PuppetDB' do
     puppetdb_client.expects(:facts_for_node).with(facts.keys).returns(facts)
-    result = run_plan('puppetdb_fact', 'nodes' => facts.keys)
+    result = run_plan('puppetdb_fact', 'targets' => facts.keys)
     expect(result).to eq(Bolt::PlanResult.new(facts, 'success'))
   end
 
   it 'adds facts to Targets' do
     puppetdb_client.expects(:facts_for_node).with(facts.keys).returns(facts)
-    run_plan('puppetdb_fact', 'nodes' => facts.keys)
+    run_plan('puppetdb_fact', 'targets' => facts.keys)
     facts.each do |k, v|
       expect(inventory.facts(Bolt::Target.new(k))).to eq(v)
     end
@@ -31,7 +31,7 @@ describe 'puppetdb_fact' do
 
   it 'returns an empty hash for an empty list' do
     puppetdb_client.expects(:facts_for_node).with([]).returns({})
-    result = run_plan('puppetdb_fact', 'nodes' => [])
+    result = run_plan('puppetdb_fact', 'targets' => [])
     expect(result).to eq(Bolt::PlanResult.new({}, 'success'))
   end
 end