+++ spec/integration/plugin/module_spec.rb
@@ -281,7 +281,7 @@ describe 'using module based plugins' do
         result = run_cli_json(['plan', 'run', 'test_plan', '--boltdir', boltdir], rescue_exec: true)
 
         expect(result).to include('kind' => "bolt/run-failure")
-        expect(result['msg']).to match(/Plan aborted: apply_prep failed on 1 nodes/)
+        expect(result['msg']).to match(/Plan aborted: apply_prep failed on 1 target/)
         expect(result['details']['result_set'][0]['result']['_error']['msg']).to match(
           /Plugin identity does not support puppet_library/
         )
@@ -301,7 +301,7 @@ describe 'using module based plugins' do
         result = run_cli_json(['plan', 'run', 'test_plan', '--boltdir', boltdir], rescue_exec: true)
 
         expect(result).to include('kind' => "bolt/run-failure")
-        expect(result['msg']).to match(/Plan aborted: apply_prep failed on 1 nodes/)
+        expect(result['msg']).to match(/Plan aborted: apply_prep failed on 1 target/)
         expect(result['details']['result_set'][0]['result']['_error']['msg']).to match(/Unknown plugin:/)
       end
     end
@@ -319,7 +319,7 @@ describe 'using module based plugins' do
         result = run_cli_json(['plan', 'run', 'test_plan', '--boltdir', boltdir], rescue_exec: true)
 
         expect(result).to include('kind' => "bolt/run-failure")
-        expect(result['msg']).to match(/Plan aborted: apply_prep failed on 1 nodes/)
+        expect(result['msg']).to match(/Plan aborted: apply_prep failed on 1 target/)
         expect(result['details']['result_set'][0]['result']['_error']['msg']).to match(
           /The task failed with exit code 1/
         )